[Behat][Admin] Use element instead of trait for promotion form
