+++ src/Sylius/Behat/Element/Admin/Promotion/FormElementInterface.php
@@ -16,4 +16,22 @@ namespace Sylius\Behat\Element\Admin\Promotion;
 interface FormElementInterface
 {
     public function prioritizeIt(?int $priority): void;
+
+    public function setStartsAt(\DateTimeInterface $dateTime): void;
+
+    public function setEndsAt(\DateTimeInterface $dateTime): void;
+
+    public function setUsageLimit(int $limit): void;
+
+    public function makeExclusive(): void;
+
+    public function makeNotAppliesToDiscountedItem(): void;
+
+    public function makeCouponBased(): void;
+
+    public function checkChannel(string $name): void;
+
+    public function setLabel(string $label, string $localeCode): void;
+
+    public function hasLabel(string $label, string $localeCode): bool;
 }
